Allow resource requests in manifest, fixes #66
Prior to this change resource specs in the manifest file could only
specify limits, or upper bounds. This change adds support for
'cpu_request' and 'memory_request" which specifies requests, or a lower
bound - e.g. my app needs at least this much CPU/memory to function.

It seems Nomad does not support specifying lower bounds so in the Nomad
case the lower bound is ignored. Kubernetes (ostensibly) supports it and
so it uses the requests and limits appropriately.

Here is a quick breakdown of how resources are requested from Kubernetes
depending on what is provided in the manifest (note it is also designed
with backwards compatibility in mind):

- If a request is specified, a limit must be explicitly provided
- If request and limit are both specified, request must be <= limit
- If only limit is specified, request = limit (this is what Kubernetes
  does anyways at the time of implementation, but the impelmentation
  makes it explicit)
- If neither are specified, both request and limit are set to the same
  (and original) default value of 0.5 CPU and 512M memory