Implement auth for kubectl
If Nelson is deployed in-cluster, kubectl ostensibly automagically
detects this and does the Right Thing TM. If not, we require operators
to provide a subconfig where key is the datacenter name and the
value is the path to the kubectl config file.

We do this instead of having one big config file because as far as I can
tell kubectl does not let you specify the cluster on the command line -
instead you must mutably set your cluster or context. This becomes
problematic if you have say, 2+ launch operations going at the same
time.. classic example of global mutable state and concurrency.

With the config-per-DC approach, we can shell out to kubectl and simply
change the KUBECONFIG environment variable with each call.

With this change, because we are now requiring operators to setup
the kubectl config themselves, this leads to other related changes in
this commit:
- do not require providing a cert and token - it would be tedious to
  require this when we're already asking for a kubectl config file
- as a consequence, remove the Kubernetes API client
- as a consequence of *that*, we need to re-implement the summary and
  health check operations (currently stubbed)

- also properly implement fallback for deleting workloads in K8s
  backend