Implement summaries and status with kubectl
Thankfully kubectl can dump output as JSON with a similar/same
"schema" as the API so I could reuse all my previous JSON parsing
logic.

This design assumes deployed objects are labeled with the stack
name which is a bit descriptive, not sure if there's a way around
this at the moment.